[ui] Add separate search page
The index page is now a placeholder for an upcoming dashboard page
